requirements.txt
validate_email
PyJWT
GoogleAppEngineCloudStorageClient
lxml==2.3.5
